Original change: https://googleplex-android-review.googlesource.com/c/platform/packages/modules/OnDevicePersonalization/+/25333616

+++ federatedcompute/src/com/android/federatedcompute/services/encryption/FederatedComputeEncryptionKeyManager.java
@@ -45,6 +45,7 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
 import java.util.concurrent.ExecutorService;
+import java.util.concurrent.TimeUnit;
 
 /** Class to manage key fetch. */
 public class FederatedComputeEncryptionKeyManager {
@@ -138,7 +139,7 @@ public class FederatedComputeEncryptionKeyManager {
      * deletes expired keys
      */
     public FluentFuture<List<FederatedComputeEncryptionKey>> fetchAndPersistActiveKeys(
-            @FederatedComputeEncryptionKey.KeyType int keyType) {
+            @FederatedComputeEncryptionKey.KeyType int keyType, boolean isScheduledJob) {
         String fetchUri = mFlags.getEncryptionKeyFetchUri();
         if (fetchUri == null) {
             throw new IllegalArgumentException("Uri to fetch active encryption keys is null");
@@ -162,7 +163,11 @@ public class FederatedComputeEncryptionKeyManager {
                 .transform(
                         result -> {
                             result.forEach(mEncryptionKeyDao::insertEncryptionKey);
-                            mEncryptionKeyDao.deleteExpiredKeys();
+                            if (isScheduledJob) {
+                                // When the job is a background scheduled job, delete the expired
+                                // keys, otherwise, only fetch from the key server.
+                                mEncryptionKeyDao.deleteExpiredKeys();
+                            }
                             return result;
                         },
                         mBackgroundExecutor); // TODO: Add timeout controlled by Ph flags
@@ -280,5 +285,35 @@ public class FederatedComputeEncryptionKeyManager {
             return 0;
         }
         return maxAge - cachedAge;
+    }
+
+    /** Get active keys, if there is no active key, then force a fetch from the key service.
+     * In the case of key fetching from the key service, the http call
+     * is executed on a BlockingExecutor.
+     * @return The list of active keys.
+     */
+    public List<FederatedComputeEncryptionKey> getOrFetchActiveKeys(int keyType, int keyCount,
+                                                                    int retryCount) {
+        List<FederatedComputeEncryptionKey> activeKeys = mEncryptionKeyDao
+                .getLatestExpiryNKeys(keyCount);
+        if (activeKeys.size() > 0) {
+            return activeKeys;
+        }
+        while (retryCount > 0)  {
+            try {
+                var fetchedKeysUnused = fetchAndPersistActiveKeys(keyType,
+                        /* isScheduledJob= */ false).get(1, TimeUnit.SECONDS);
+                activeKeys = mEncryptionKeyDao.getLatestExpiryNKeys(keyCount);
+                if (activeKeys.size() > 0) {
+                    return activeKeys;
+                }
+            } catch (Exception e) {
+                LogUtil.e(TAG, "Exception encountered when forcing encryption key fetch: "
+                        + e.getMessage());
+            } finally {
+                retryCount -= 1;
+            }
+        }
+        return activeKeys;
     }
 }
